brokers/sinotrade-stock-python/requirements.txt
shioaji==1.3.1
requests==2.32.5
python-dotenv==1.2.1
py-spy==0.4.1
